work/s2_lookup/requirements.txt
requests
python-dateutil
nltk
orjson
